demo/demo-ua/src/main/java/email/EmailValidation.java
package email;

import jakarta.mail.internet.AddressException;
import jakarta.mail.internet.InternetAddress;
import org.apache.commons.validator.routines.EmailValidator;

public class EmailValidation {

    public boolean isValidWithRegex(String email) {
        return email.matches(".*@+.*\\..*");
    }

    public boolean isValidWithRegex02(String email) {
        String regex = "^[a-zA-Z0-9_+&*-]+(?:\\.[a-zA-Z0-9_+&*-]+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,7}$";
        return email.matches(regex);
    }

    public boolean isValidWithJakartaMail(String email) {
        try {
            InternetAddress emailAddr = new InternetAddress(email);
            emailAddr.validate();
            return true;
        } catch (AddressException e) {
            return false;
        }
    }

    public boolean isValidWithApacheCommons(String email) {
        EmailValidator emailValidator = EmailValidator.getInstance();
        return emailValidator.isValid(email);
    }

}
